Move logic to write table part to TablePartWriter. It decreases the clutter and hopefully increases encapsulation, so one writer can optimize how to write its own part without another 4k+ LOC.
